Avoid modifying the original klipper.service
We can use systemd's drop-in support to add a `ExecStartPre` step without modifying the original file.  This removes the risk of the `sed` approach possibly messing up the syntax of the file (although unlikely), and additionally preserves the change even if a new version of `klipper.service` is shipped from upstream.